♻️ refactor: refactor a preference slice
