+++ src/store/global/selectors.ts
@@ -1,2 +1,2 @@
-export * from './slices/common/selectors';
+export * from './slices/preference/selectors';
 export * from './slices/settings/selectors';
