+++ src/store/global/slices/common/action.ts
@@ -1,4 +1,3 @@
-import { produce } from 'immer';
 import { gt } from 'semver';
 import useSWR, { SWRResponse } from 'swr';
 import type { StateCreator } from 'zustand/vanilla';
@@ -9,12 +8,11 @@ import { CURRENT_VERSION } from '@/const/version';
 import { globalService } from '@/services/global';
 import type { GlobalStore } from '@/store/global';
 import type { GlobalServerConfig } from '@/types/settings';
-import { merge } from '@/utils/merge';
 import { setNamespace } from '@/utils/storeDebug';
 
-import type { GlobalCommonState, GlobalPreference, Guide, SidebarTabKey } from './initialState';
+import type { SidebarTabKey } from './initialState';
 
-const n = setNamespace('settings');
+const n = setNamespace('common');
 
 /**
  * 设置操作
@@ -26,12 +24,7 @@ export interface CommonAction {
    * @param key - 选中的侧边栏选项
    */
   switchSideBar: (key: SidebarTabKey) => void;
-  toggleChatSideBar: (visible?: boolean) => void;
-  toggleExpandSessionGroup: (id: string, expand: boolean) => void;
-  toggleMobileTopic: (visible?: boolean) => void;
-  toggleSystemRole: (visible?: boolean) => void;
-  updateGuideState: (guide: Partial<Guide>) => void;
-  updatePreference: (preference: Partial<GlobalPreference>, action?: string) => void;
+
   useCheckLatestVersion: () => SWRResponse<string>;
   useFetchGlobalConfig: () => SWRResponse;
 }
@@ -47,51 +40,6 @@ export const createCommonSlice: StateCreator<
   },
   switchSideBar: (key) => {
     set({ sidebarKey: key }, false, n('switchSideBar', key));
-  },
-  toggleChatSideBar: (newValue) => {
-    const showChatSideBar =
-      typeof newValue === 'boolean' ? newValue : !get().preference.showChatSideBar;
-
-    get().updatePreference({ showChatSideBar }, n('toggleAgentPanel', newValue) as string);
-  },
-  toggleExpandSessionGroup: (id, expand) => {
-    const { preference } = get();
-    const nextExpandSessionGroup = produce(preference.expandSessionGroupKeys, (draft) => {
-      if (expand) {
-        if (draft.includes(id)) return;
-        draft.push(id);
-      } else {
-        const index = draft.indexOf(id);
-        if (index !== -1) draft.splice(index, 1);
-      }
-    });
-    get().updatePreference({ expandSessionGroupKeys: nextExpandSessionGroup });
-  },
-  toggleMobileTopic: (newValue) => {
-    const mobileShowTopic =
-      typeof newValue === 'boolean' ? newValue : !get().preference.mobileShowTopic;
-
-    get().updatePreference({ mobileShowTopic }, n('toggleMobileTopic', newValue) as string);
-  },
-  toggleSystemRole: (newValue) => {
-    const showSystemRole =
-      typeof newValue === 'boolean' ? newValue : !get().preference.mobileShowTopic;
-
-    get().updatePreference({ showSystemRole }, n('toggleMobileTopic', newValue) as string);
-  },
-  updateGuideState: (guide) => {
-    const { updatePreference } = get();
-    const nextGuide = merge(get().preference.guide, guide);
-    updatePreference({ guide: nextGuide });
-  },
-  updatePreference: (preference, action) => {
-    set(
-      produce((draft: GlobalCommonState) => {
-        draft.preference = merge(draft.preference, preference);
-      }),
-      false,
-      action,
-    );
   },
   useCheckLatestVersion: () =>
     useSWR('checkLatestVersion', globalService.getLatestVersion, {
